+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Fix FormulaWidget column check [#437](https://github.com/CartoDB/carto-react/pull/437)
+
 ## 1.3
 
 ### 1.3.0-beta.2 (2022-06-20)