+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Fix CategoryWidget search if there are null values [#439](https://github.com/CartoDB/carto-react/pull/439)
 - Layout improvements in BarWidgetUI [#438](https://github.com/CartoDB/carto-react/pull/438)
 - Fix FormulaWidget column check [#437](https://github.com/CartoDB/carto-react/pull/437)